helper function in module exports
